Close the code blocks and puts thread titles
The thread title thing might be replaced by
actual thread_prompt system, because this
looks more natural to me.

The code blocks closing code was generated and
is... kinda awful actually. I should rework it
if I have time, but this bot is just a tool
for me at the moment.

Signed-off-by: Voyage <voyage@miouyouyou.fr>